feat: PcodeBackend shared state + hexagon pcmgr
Pcodebackend has shared state in the form of an fxhashmap that maps from
a key in an enum to a u128. Used to share info about packet starts and
ends between the GeneratorHelper and the PCManager.

The PC manager now makes sure to set the ISA PC in the RegisterManager
and is updated at packet granularity or on a branching instruction.
Branching instructions are yet to be tested.