feat: add pcode cache for faster translating
Adds a cache of address -> pcode in the PcodeTranslator. Translates are
checked for a cache hit and the 16 bytes starting from the address.

Currently yielding ~60% faster emulation for the pcode fib register
benchmark. Unfortunately there is a ~10% performance penalty if caching
is disabled due to the pcode cloning required. The perf hit from cloning
is offset by caching if enabled.

With the freertos test I see ~46% performance improvement. I think this
is because less instructions are able to be cached.